[TASK] Add sub navigation only when there are pages on treeLevel 1
We define a sub navigation with TypoScript. It is
basically the same as the main navigation, except
that we define the `entryLevel` for the menu.

In FLUID we check if there are sub navigation with
a condition. When there are items rendered, the necessary HTML markup is also added around the
sub navigation, which is generated with TypoScript.