[TASK] Sujet image on subpage(s)
The editor of the website add an image to a page
(edit page properties).

The markup for the sujet image and the surrounding
div is generated with TypoScript (lib.sujet).
The TypoScript tries to grab the image from the
actual page. If there is no image, it tries to grab
the image from a page up the rootline.
When an image is found, it gets automatically wrapped
by TypoScript.

We use the f:cObject viewhelper to include it in
the FLUID template.